requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.2.0
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
# langchain-google-genai>=2.0.0  # desactivado - usar Groq + Ollama
pydantic-settings>=2.0.0
chromadb>=0.5.0
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.0.0
